Documentation of manifest file and global sort of rects
Global sort of rects means in a multi-threaded situation, the
requests that end up in each thread generally are for the same
area of the sky. This means when the server unzips a patch there
is a greater likelihood the same request will contain all sources
in that patch for the entire download.