Include all immutable dc.Rect fields in manifest so Resume works in all configs.
This is a breaking change for the manifest file format, in that downloads
started before this change will not be able to resume if resumed using a version
of fibad after this change.

Prior to this change If you have a config that downloads mask or variance, resume
would simply fail (issue #98)

The cause was that the image, variance, and mask rect parameters weren't saved to the
manifest file, so on resume _prune_downloaded_rects() could not match the  manifest
with the rects generated from the catalog.

In this case the rects generated from the catalog had config-provided values for
image, mask and variance. The rects generated from reading the manifest had the
dc.Rect default values of image=True, mask=False and variance=False.

This change ensures that all the dc.Rect immutable fields, which are used in
equality and hash checking are written out to the manifest.