First implementation of cache warming for HSCDataSet (#245)
This spawns a pre-cache thread during HSCDataSet initialization. The pre-cache thread starts many worker threads that read in all the dataset files to the tensor cache. The idea is to confront a latency-limited network storage backend with a large number of requests all at once, such that we will receive the files in memory at the greatest overall bandwidth.

This change also logs a HSCDataSet/preload_1k_obj_s metric to tensorboard which should give an idea of object bandwidth. It is the elapsed time to load 1000 objects measured every 1000 objects.

The change is protected under a config config["data_set"]["preload_cache"] The config is on by default (similar to use_cache) because it should be harmless for small datasets and laptop usage.

We have hardcoded the correct number of processes on Hyak